# ConsoleApp1/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NetStore.Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Orders/GetOrder/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get Order By Id returns List ResponceOrderDetailDto",
        "operationId": "GetOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Orders": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Post Order (Request Order)",
        "operationId": "PostOrder",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "FullName": {
                    "type": "string"
                  },
                  "Address": {
                    "type": "string"
                  },
                  "Phone": {
                    "type": "string"
                  },
                  "OrderTotal": {
                    "type": "number",
                    "format": "double"
                  },
                  "OrderPlaced": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "IsOrderCompleted": {
                    "type": "boolean"
                  },
                  "UserId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "FullName": {
                  "style": "form"
                },
                "Address": {
                  "style": "form"
                },
                "Phone": {
                  "style": "form"
                },
                "OrderTotal": {
                  "style": "form"
                },
                "OrderPlaced": {
                  "style": "form"
                },
                "IsOrderCompleted": {
                  "style": "form"
                },
                "UserId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Orders/CompletedOrders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get All Completed Orders",
        "operationId": "CompletedOrders",
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Orders/PendingOrders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get All Pending Orders",
        "operationId": "PendingOrders",
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Orders/CountOrders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Count All Orders",
        "operationId": "CountOrders",
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Orders/OrderByUser/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get Orders By User",
        "operationId": "OrderByUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/Orders/MarkOrderCompleted/{id}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "Mark Orders as Completed",
        "operationId": "MarkOrderCompleted",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}